# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ant-design v5 接口demo",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/admin_users/ids_destroy": {
            "delete": {
                "tags": [
                    "用户"
                ],
                "summary": "删除用户",
                "description": "删除用户",
                "operationId": "App\\Http\\Controllers\\AdminUsersController::destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/permissions/ids_destroy": {
            "delete": {
                "tags": [
                    "权限"
                ],
                "summary": "删除权限",
                "description": "删除权限",
                "operationId": "App\\Http\\Controllers\\PermissionsController::destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "权限id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/roles": {
            "get": {
                "tags": [
                    "角色"
                ],
                "summary": "角色列表",
                "description": "角色列表",
                "operationId": "App\\Http\\Controllers\\RolesController::index",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "角色",
                        "schema": {
                            "type": "string"
                        },
                        "example": "超级管理员"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "角色列表"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "角色"
                ],
                "summary": "添加角色",
                "description": "添加角色",
                "operationId": "App\\Http\\Controllers\\RolesController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "角色名",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permission_ids",
                        "in": "query",
                        "description": "权限ids",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功"
                    },
                    "422": {
                        "description": "错误的凭证响应"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/roles/{role}": {
            "patch": {
                "tags": [
                    "角色"
                ],
                "summary": "修改角色",
                "description": "修改角色",
                "operationId": "App\\Http\\Controllers\\RolesController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "角色id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "角色名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permission_ids",
                        "in": "query",
                        "description": "权限ids",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/roles/ids_destroy": {
            "delete": {
                "tags": [
                    "角色"
                ],
                "summary": "删除角色",
                "description": "删除角色",
                "operationId": "App\\Http\\Controllers\\RolesController::destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "角色id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}